Devel: Remove coccinelle support.
This has seen one commit in the last four-ish years, so clearly no one
is too excited about using it.  And the reason for that is it's just
extremely difficult to write the semantic patches - it's in a weird
custom language, documentation is terse and kind of hard to follow, and
personally I never had a lot of confidence that it would catch
everything.