Test: tap-driver.sh:  Update from the latest in automake source.
We've been having CI problems that seem to originate from gawk being
called by tap-driver.sh trying to allocate way too much memory, but only
on certain platforms.  As a first step, let's update this script (which,
we were using a version from 2011) and see if that helps anything.  If
not, it's at least nice to update this ancient code.