Doc: Fix top-level index.rst files.
Prior to this patch, in the PDF documentation, the abstract was chapter
1, the table of contents was chapter 2, and the index was chapter 3.
All the content of the book was in subheadings under chapter 2, which is
not really what we're looking for.

Now, the abstract is no longer its own chapter - it's just some text on
its own page following the table of contents - and the actual book
follows in its own chapters.