Refactor: Standardize vim modelines in .in files.
If it's a .in file that vim has syntax highlighting support for, add a
modeline to the bottom of the file that sets the filetype.

If there was already a modeline, reduce it down to just setting the
filetype.  Other settings are global and should either be in a top-level
vimrc file (which we don't have, and I don't know if vim supports but
still) or in personal config files.