Refactor: daemons: Use pcmk__request_t for LRMD_OP_RSC_EXEC.
Additionally, execd_process_rsc_exec should return a standard pacemaker
return code.  It doesn't do anything useful with the call_id value it
extracts and returns, and the caller already knows the call_id (and
doesn't do anything with the return value from execd_process_rsc_exec
aside from check if it's an error or not).