Refactor: daemons: Use pcmk__request_t for LRMD_OP_RSC_UNREG.
This requires a little extra work, in addition to the steps of exposing
a function and changing its return type to be a standard Pacemaker
return code.

Sending a generic notify for unregistering a resource also requires
inspecting the return value, which we luckily have in the reply XML so
we can just grab it from there as needed.