Refactor: daemons: Add execd_process_message.
At the moment, this function is fairly unfortunately named (there's
already a process_lrmd_message).  However, the new function is gradually
going to be taking code from the old function until the old one is
completely gone.  So it's just a temporary confusion we'll need to live
with for a bit.

This new function holds the framework for handling commands with
pcmk__request_t instead of the old system.  It's a little tricky to
convert these commands over one at a time and still keep everything
working, which is why there's the big TODO comment.