Refactor: daemons: Condense duplicated IPC code.
Both places that call execd_process_message first add several additional
attributes to the request XML, so it makes sense to move that code into
execd_process_message instead.

Note that there are two different approaches to setting client->name if
it's missing.  I don't see why the attribute would ever be missing, so I
think it's safe to get rid of the lengthier approach that grabs the
client PID as a last option.