Low: libcrmcommon: Keep trailing newlines in pcmk__scan_nvpair()
The current callers won't have trailing newlines from the CLI or from
reading from a file. Newlines would be present only if the user included
them explicitly.

Newline removal was (re-)added by b7fe638, to fix a regression in
28ca6aa. However, at the time of 28ca6aa, pcmk_scan_nvpair() was public
API, so the change in behavior actually mattered. Now, it doesn't. It
appears to be safe to stop stripping trailing newlines.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>