Low: libcrmcommon: Allow empty string values in pcmk__scan_nvpair()
I don't see any reason not to. This does change behavior for
crm_resource and stonith_admin if "NAME=" is given as a command line
argument. However, this is such an edge case that I don't expect it to
matter. It seems foolish to depend on the current behavior.

Of course, that means nothing, so if anyone complains, we can revert.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>