Low: daemons: Send a NACK instead of an ACK when there's an error.
I don't know the reasoning behind this, but it seems strange for daemons
to send an ACK along with an error code.  For starters, it seems a
little redundant to have both an ACK vs. NACK, as well as a status code.

But since we have that, it's confusing to send both an ACK (which seems
like a positive response) and an error code (which is obviously not).

So, anywhere we'd respond with an error code (aside from
CRM_EX_INDETERMINATE which means we don't know yet), send a NACK
instead.