Feature: tools: Don't require --force for non-privileged cibadmin ACLs
Previously, we required the --force argument to proceed with `cibadmin
--show-access ...` when the command was run as a non-privileged user
(that is, as a user other than CRM_DAEMON_USER -- usually hacluster --
or root). However, a warning that the output may be incorrect seems
sufficient and makes the code simpler.

In the future we could cause this command to fail unconditionally if run
as a non-privileged user.

Also, clarify the warning message.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>