Log: various: Avoid warnings about a negative value for `stonith-watchdog-timeout`
Despite the discussion from
https://github.com/ClusterLabs/pacemaker/pull/3677 about some potential
ideas for the future, so far it has been documented that a negative
value for `stonith-watchdog-timeout` is valid. But with 56a00acc6a,
crm_get_msec() started explicitly warning about a "Negative" value being
invalid.

Even if we decide to forbid any negative value for
`stonith-watchdog-timeout` in the future, it'll need a proper
deprecation and eventually a transformation. So no matter how we'll
proceed, for now we probably have to parse it as an integer first to
avoid the warning from crm_get_msec() .