Refactor: tools: Add get_live_peers to cibsecret.
This is a function that returns a list of peer nodes that are pingable
and will be used by various subcommands.  Since none of those are
implemented yet, the function is currently marked as unused to preserve
compilability.

Some of this definitely duplicates code elsewhere.  For instance, this
code is largely pcmk__list_nodes, but returning a list of strings
instead of using formatted output.