Refactor: tools: Implement the sync subcommand in cibsecret.
Error handling is somewhat obnoxious in this subcommand.  I assume it's
going to be the same way in all the others.  If any of the rsh/rcp
commands failed, cibsecret.in would just exit with CRM_EX_ERROR, so
that's what we're going to do here.

These functions do return more useful codes, but there's nothing we can
do with them at the moment.