Feature: tools: cibadmin ignores --node-path for non-XPath-query command
The output printing logic didn't make sense for the output of other
commands. It made a recursive call under the assumption that the
top-level output element was PCMK__XE_XPATH_QUERY.

Full recursion is also unnecessary. Based on the structure of the
output, we only need to go one level deep, and we only need to look for
PCMK__XE_XPATH_QUERY_PATH elements.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>