API: tools: Deprecate cibadmin -N/--node option
There are no known internal or external users of this option. It appears
that it was added in 2004 as "--host" by commit 203b8e54 for no
particular reason. Perhaps it was for consistency with other tools, or
perhaps we thought it was a good idea to expose as much internal
functionality as possible.

I can't think of a good use case for directing a cibadmin command to a
particular node, unless the cluster is in a very odd state. At that
point, manual intervention on that node would be reasonable.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>