Refactor: tools: New cibadmin flags for requiring input
Previously, we might call read_input() even if we didn't need input. As
of this commit, we call it only if we need the input. That's the reason
for the return code change here.

The "Please supply XML to process" message in output_digest() goes away
because we handle it more generally in the new block (if (!as_xpath)).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>