API: tools: Deprecate cibadmin -e/--node-path option
There are no known internal or external users of this option. It was
added in 2012 by commit dd94186 without any reason given. There may be
some use case, but I struggle to think of one that isn't contrived.

If no one needs this feature, then it's clutter and an unnecessary
maintenance hassle. We'll plan to remove it in a future release unless
someone needs it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>