API: libcib: Deprecate cib_xpath_address
There are no known external users of this option. The only internal user
is cibadmin.c, where the code that uses it is deprecated.

It was added in 2012 by commit dd94186 without any reason given. There
may be some use case, but I struggle to think of one that isn't
contrived.

If no one needs this feature, then it's clutter and an unnecessary
maintenance hassle. We'll plan to remove it in a future release unless
someone needs it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>