Test: python: Add a new cibsecret ctslab test.
This does not test every aspect of the current cibsecret program, but it
does at least test every subcommand which covers most of how it will be
used.  Most notably, it does not test pssh and pdsh.

I can't remember when we last added a new ctslab test - this one doesn't
seem completely in the spirit of ctslab but it's the only automated
testing we've got that involves nodes running.