Low: libcrmcommon: Return a value for md5sum of empty input.
g_compute_checksum_for_data will return NULL, but there are times where
we may want the md5sum of an empty string - for example, in cibsecret.c.
Luckily the md5sum of an empty string is always the same, so we can just
hardcode it.

I don't think this is worth adding a unit test for since it's just a
hardcoded string.