API: schemas: Add a schema for cibadmin
We can have no output, an "md5-sum" element, or an "output" element
containing CData.

The output element contains a CData block that is typically a
stringified version of an XML tree. The one exception occurs when
--show-access is used either with no value or with a value of "color" or
"text". In those cases, the output contains pieces of an XML tree but
also contains some invalid characters that are meant to make the output
more human-readable.

An alternative would be to use a different output element (for example,
named "rendered-acls" instead of "output") for --show-access queries.

Closes T116

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>