Log: libcrmcommon: Drop crm_perror() in pcmk__write_series_sequence()
A similar error in the next block already uses crm_err(). fprintf() does
not set errno, so it makes no sense to print it. Printing to stderr
could be useful for CLI tools that use the cib_file interface. But CLI
tools do have verbose flags. So it doesn't seem worth the hassle of
passing a pcmk__output_t object down the call chain.

Ref T172

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>