Log: libcrmcommon: Drop crm_perror() in mainloop.c
The error message in mainloop_destroy_signal() is made redundant by the
error messages in crm_signal_handler().

The only possible benefit of crm_perror() in these two cases is that
it writes the messages to stderr for CLI tools that use a mainloop. That
should be a rare situation, and no other mainloop errors get written to
stderr.

CLI tools have verbose flags. So it doesn't seem worth the hassle of
passing a pcmk__output_t object down the call chain.

Ref T172

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>