Log: libcrmcommon: Drop crm_perror() in xml_io.c
Note that fprintf() doesn't set errno on error.

The only added benefit of crm_perror() in these cases is that it writes
the messages to stderr for CLI tools that read and write XML. However,
no other XML I/O errors get written to stderr, and CLI tools have
verbose flags. So it doesn't seem worth the hassle of passing a
pcmk__output_t object down the call chain.

Ref T172

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>