Log: libcrmcommon: Drop crm_perror() in logging.c
And improve log messages. Note that errno should be set to 0 before
calling getpwuid(), and that it does not necessarily set errno if no
matching entry is found.

Both of these calls are for daemons. We don't typically see stderr from
daemons. So we're not really losing anything by not writing these
messages to stderr as crm_perror() did.

Ref T172

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>