Log: libcrmcommon: Drop crm_perror() in digest.c
If calculated == NULL here, then either we already logged an assertion
via CRM_CHECK(), or there was an error in GLib, in which case GLib
probably logged an assertion. In any case, we can't expect errno to be
set to a meaningful value, so there's no reason to log it as
crm_perror() does.

We lose logging to stderr for CLI tools that use the cib_file interface.
But this should be a rare situation, and CLI tools have verbose flags.

Ref T172

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>