Refactor: libcrmcommon, tools: Clean up crm_mon on failure to daemonize
This arguably makes pcmk__daemonize() a little bit less intuitive, but I
didn't like passing in a pointer to crm_mon.c's clean_up() function and
passing the exit code into it.

This is so that we can free data structures and write any output from
the pcmk__output_t object.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>