Doc: ChangeLog: update for 3.0.1 release
Note:  The only code difference between this and the last changelog
update is the ECONNREFUSED fix.  That fixes a regression introduced
after 3.0.0.  It's not mentioned anywhere in the changelog because it's
just an update to another fix and was not in a final release anywhere.

Thus, this patch just modifies the changelog summary.