Doc: cts-cli: Drop FIXME comment about cibadmin digests
In 948c160, I carelessly removed the sed command that stripped the
digest. It was added by 10542c49 to resolve some digest mismatch issues
that were observed. However, that commit message says that we haven't
observed any issues locally or on ClusterLabs CI systems. This raises
the question "Where DOES the issue occur?"

If anyone complains, we can strip out the digest again.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>