Refactor: libpe_status: Drop uses of "orphan" where possible
A lot of these replacements are pretty clear.

It is unclear to me, however, whether orphan clone instances should be
referred to as "removed." This is why I put the word "removed" in quotes
in some of the log messages and comments. I don't fully understand how
they end up existing in the first place, but "removed" seems like it
might be misleading. It's probably good enough for now.

Ref T894

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>