Test: cts-scheduler: Use "fencing-timeout" in inputs
Add comments in five tests where stonith-timeout is used as an instance
attribute for fencing resources. We can't drop or rename the attributes,
because this causes a digest mismatch, which causes a restart due to
resource definition change.

Pacemaker and probably the fencing agents ignore the stonith-timeout
instance attribute for fencing resources. But it affects digests.

Ref T895

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>