Refactor: controller: Case-insensitive hash table for fencing failures
We always use case-insensitive comparisons for node names. Perhaps we
shouldn't, but that's been supported for a long time and we've made a
point to continue doing so. (I don't have a citation for this; it came
up frequently in discussions with Ken though.)

It probably makes no difference here, but we'll do it for consistency.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>