Refactor: tools: Create crm_mon.c cib object later
It's not needed until the call to one_shot(). There was probably a time
when we based decisions on the cib->variant sooner, as the comment
suggested. However, the only things that use cib->variant now are
mon_refresh_display(), setup_api_connections(), and the scheduler object
initialization. All of those happen later.

This avoids the need to think about the cib object when daemonizing.

(cib__set_output() checks cib->variant, but simply returns early if not
cib_remote.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>