Refactor: cts: Drop Environment._seed_random()
* self["RandSeed"] isn't used anywhere except a log line in cts-lab.in.
* From the random.Random.seed() documentation: "None or no argument
  seeds from current time or from an operating system specific
  randomness source if available."
* The specific default seed source is an implementation detail that
  doesn't really matter.

Perhaps keeping this for logging the random seed could be useful for
reproducing hard-to-reproduce issues that emerge in testing. FOR THE
MOST PART, I'd expect that simply re-running the same test or sequence
of tests would allow for reproducing an issue. There are some other uses
of randomness though, like the fencing configuration.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>