Test: cts: Drop -a and screen functionality from cts shell script
Apparently, the cts script opens a screen session with certain cluster
variables exported by default if no other command is given. This is not
documented in the help text. It is also not documented that the screen
command must be installed.

This has been present since the cts script was added by f17324fb.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>