Test: cts: Drop -l/--logfile option from cts script
It doesn't seem useful. It shows the log file that would be used for
summary and watch commands. This log file does NOT get used for cts-lab
runs via the cts script.

Further, it's always just "~/cluster-$cluster_name.log" unless the user
specified a log file on THIS command line using either the -f/-sf option
or a custom cluster name. (Note that if a log file name is set when this
script creates a custom cluster, that name is not saved anywhere for
reuse. Future uses of the cts script with the same cluster name use the
default log file name. See the "case $cluster_name in" block.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>