Test: cts: Drop -x option from cts script
If we want to turn on bash tracing, we should do so when we're invoking
the script. Not within the script itself. The way it was set up, the -x
option did not take effect until the script's option parser reached that
option. So tracing would be enabled part of the way through the script.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>