Test: cts: Drop -s option from cts script
The below assumes that the -f/-sf option was not used. -f/-sf implies -s
but overrides the hard-coded default log file with a custom one.

The -s (summary) option uses the cts script as a glorified grep wrapper.
It runs `grep` on a log file that's hard coded based on the provided
cluster name (or based on the custom cluster creation prompts from
earlier in this command). It greps for lines containing "CTS:" and any
other patterns that the user provides with the "-e" option. It excludes
lines with "debug:" and "cmd:" based on the -v (verbosity) option.

These are things that a user can easily do on their own without the use
of this script. This is especially true since the hard-coded log file
has no reason to exist unless the user had read this script and passed
that log file name to their cts-lab run.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>