Test: cts: Drop -t and positional test number options from cts script
This is one of the most useful parts of the cts script, but IMO not
useful enough to keep. I could be convinced otherwise, however.

This takes a list of test numbers (and ranges of test numbers) from a
cts-lab execution log (the cts-lab --outputfile argument).

For each test number or range, it runs crm_report to generate an output
directory.
* It checks either the cts script's default log file (which, as
  mentioned in previous commits, does not generally exist) for this
  cluster, or the file specified by the cts -f/-sf option.
* It gets the test's start and end timestamps (or the start timestamp of
  the first test in the range and the end timestamp of the last test)
  from that log file.
* It collects logs (from all nodes in the test cluster, or just from the
  local node? crm_report and its helper scripts are hard to read)
  between those start and end timestamps.

Again, this is potentially useful, but it doesn't seem that much easier
than simply checking the logs on each test cluster node. In practice,
cts is basically for developer use, and we rarely test clusters so large
that checking logs by hand is not feasible. Besides, we would only need
to check the system and cluster logs if a test behaves unexpectedly, and
we'd prefer to investigate that on a manageably sized cluster if
possible.

If needed, we could always run `crm_report` with appropriate start and
end timestamps ourselves.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>