Test: cts: Drop -f/-sf option from cts script
The -f/-sf option uses the cts script as a glorified grep wrapper. It
runs `grep` on a log file that's provided in the option's argument. It
greps for lines containing "CTS:" and any other patterns that the user
provides with the "-e" option. It excludes lines with "debug:" and
"cmd:" based on the -v (verbosity) option.

These are things that a user can easily do on their own without the use
of this script.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>