Test: cts: Fix regression in ComponentFail test
Commit a58a4000 caused a regression such that the ComponentFail test
always chooses to kill corosync. It never kills any of the other
Pacemaker daemons, which all get filtered out because they're in
valgrind-procs.

Prior to that commit, they got filtered out only if valgrind-tests was
set. It defaulted to False. The 'if self.env["valgrind-tests"]' check
was removed by that commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>