Refactor: cts: Drop key == "Name" test from ClusterManager.__getitem__()
name is a public member, so access it directly. It wouldn't hurt to
make it a property in the future, but there's no reason to treat it like
a dict member.

We replace a couple of existing self.templates["Name"] uses as well.
Note that self.templates["Name"] ends up being exactly self.name for a
Corosync2 instance ("crm-corosync") and incorrect for any other type of
ClusterManager ("crm-base").

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>