Refactor: cts: Reorganize common-ignore patterns
This way get_component("X-ignore") returns all "ignore" templates for a
given component "X".

A better solution might be to have each patterns component consist of
two lists (one to ignore and one not to ignore). We can come back to
that later perhaps. I mostly wanted to get this out of cmcorosync.py.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>