Low: sysconfig: Remove --leak-check=full from default VALGRIND_OPTS.
Both the valgrind and callgrind tools read this environment variable.
However, callgrind does not support --leak-check=.  It does support all
the other options we enable by default.  If it sees an option it doesn't
understand, it prints an error message and quits.  This means that if
you enable callgrind and don't change our default VALGRIND_OPTS, you'll
get a bunch of errors in the pacemaker log about daemons not starting,
but the reason why is undiscoverable due to closing file descriptors
when we fork off the subdaemons.

I considered two other ways of fixing this:

(1) Filtering out --leak-check= from the environment variable.  However,
there are probably other options that callgrind does not support and I
don't want to constantly have to chase that down.  Further, I think our
default setting should work with all the valgrind-related tools that we
expose a flag for.

(2) Having two settings in sysconfig and choosing the right one
depending on which tool is enabled.  This seems unnecessarily confusing.

Fixes T988