Refactor: cts: Drop --experimental-tests option from cts-lab
SplitBrainTest has been marked as experimental since 2009. There are no
other experimental tests, and this option is False by default. So
SplitBrainTest has been excluded from most of our cts-lab runs.

The experimental flag was set on SplitBrainTest by commit 84d30e0. As of
that commit, the "is_unsafe" flag already existed. "Unsafe" tests are
included by default but can be excluded by the "--no-unsafe-tests"
option. "Experimental" tests are excluded by default but can be included
by the "--experimental-tests" option.

That commit message doesn't give any suggestion of why SplitBrainTest
was marked experimental instead of unsafe. The message says only "A
bunch of cleanups so that we dont peform inappropriate test when ocfs2
is active". That seems like the intended use case for the is_unsafe
flag:

  --no-unsafe-tests     Don't run tests that are unsafe for use with
                        ocfs2/drbd

In any case, if a user with ocfs2 was concerned about the safety of
their cluster, they were probably passing --no-unsafe-tests and NOT
passing --experimental-tests. (Supposing for a moment that there were
any external users.) So this should result in no practical changes.

P.S. When the --experimental option itself was added -- NOT when the
is_experimental flag was set for SplitBrainTest -- the --no-unsafe-tests
option and the is_unsafe flag did not yet exist. See 0eb08ff3.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>