Refactor: cts: Drop --oprofile option from cts-lab
Performance profiling could be useful! For example, using the perf
utility or even the latest OProfile's operf tool.

However, Fedora 42 does not ship OProfile. I'm not sure when it stopped
being shipped, or what other distros do or do not ship it. The latest
upstream release was version 1.4.0 in 2020.

More importantly, all of cts-lab's uses of OProfile are based on the
opcontrol command. This command was dropped in OProfile version 1.0.0,
released in 2014. So this cts-lab option would not work on any remotely
modern system without some work to install a very old version of
OProfile.

https://oprofile.sourceforge.io/news/
https://oprofile.sourceforge.io/release-notes/oprofile-1.0.0

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>