Refactor: cts: Drop --warn-inactive option from cts-lab
It's been disabled by default since 2008 (see commit 62f47561). As
usual, that commit message gives no explanation of why that decision was
made. However, a resource may be inactive for any number of reasons. For
example, a constraint may allow the resource to run only on a particular
node, and that node may be inactive.

It seems better not to warn and alarm the user in general. Whether the
resource is EXPECTED to be inactive, depends on the situation or the
test being run.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>