Refactor: cts: Drop --fencing/--stonith option from cts-lab
This was a moderate source of confusion fo me when I first started using
cts-lab. We have a --fencing (or --stonith) option, as well as
--stonith-type and --stonith-args options.

The --stonith-type and --stonith-args options have their own default
values. The --fencing option may or may not override those defaults.
Then if the --stonith-type or --stonith-args options are given, they
override anything that was set by the --fencing option.

The --fencing option is a convenient shorthand for some common fencing
configurations, but we don't need it. A user can configure values
explicitly using --stonith-type and --stonith-args.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>