Refactor: controller: Drop old_len from register_fsa_input_adv()
If old_len == 0, then it doesn't matter whether we prepend or append to
the empty list. So we can set prepend to FALSE unconditionally.

We don't need to dump the queue here. Since we've stalled the FSA, we
should exit the FSA soon within s_crmd_fsa(). We dump the queue there.

It should be impossible to fail to add the message to the queue; GLib
should assert. In any case, we assume g_list_prepend() and friends work
everywhere else.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>